Enable org export
Since org mode won't generate a TOC within itself (and it shouldn't, that's the point of headers and folding AFAIK), maybe exporting from org to org might do the trick. The TOC is useful not much for displaying the org file itself but when you use it in other places like github where you can't fold headers and have a TOC at hand, that's when a TOC comes in handy.